Variation Jam
This is a game about collecting 10 pellets. It has three game modes:
- The Rush: A ghost pursues the player while they attempt to collect all 10 pellets.
- The Double: Similar to The Rush, but with two ghosts chasing the player.
- The Hunt: A straightforward game mode where the pellets randomly spawn, and the player collects them without any ghosts.

I was inspired to create this game because of Pac-Man. I’ve always enjoyed the challenge of avoiding ghosts while collecting items, and I wanted to bring that fun experience into a new game with different modes and challenges.